Add a new routine to detect and print pins and pinning moves
The definition of pin used is specified in pin.c. See the code for detailed comments.

It is important to note that the pin may be trivial to defend by the pinned side,
e.g. by capturing a hanging attacker. That's why some absurd-looking moves
may appear in the list of pinning moves: they do create a pin, but not an useful one.